enemy waves do not work yet

+++ MergeSort.cpp
@@ -15,10 +15,10 @@ void MergeSort<T>::Init(dynamic_array<T>* lis) {
 	list = lis;
 }
 template<class T>
-void MergeSort<T>::Sort(int l, int h)
+void MergeSort<T>::Sort(size_t l, size_t h)
 {
 	if (l < h) {
-		int mid = (l + h) / 2;
+		size_t mid = (l + h) / 2;
 		Sort(l, mid);
 		Sort(mid + 1, h);
 		Merge(l, mid, h);
@@ -28,9 +28,9 @@ void MergeSort<T>::Sort(int l, int h)
 
 
 template<class T>
-void MergeSort<T>::Merge(int l, int mid, int h)
+void MergeSort<T>::Merge(size_t l, size_t mid, size_t h)
 {
-	int i = l, j = mid + 1;
+	size_t i = l, j = mid + 1;
 	dynamic_array<T> A(16);
 	while (i <= mid && j <= h) {
 		if (list->get(i) < list->get(j))
@@ -48,8 +48,8 @@ void MergeSort<T>::Merge(int l, int mid, int h)
 		A.push_back(list->get(j));
 	}
 	//copy to the list
-	int k = 0;
-	for (int i = l; i <= h; i++)
+	size_t k = 0;
+	for (size_t i = l; i <= h; i++)
 	{
 		list->get(i) = A.get(k++);
 	}
@@ -58,9 +58,9 @@ void MergeSort<T>::Merge(int l, int mid, int h)
 }
 
 template<>
-void MergeSort<Collider*>::Merge(int l, int mid, int h)
+void MergeSort<Collider*>::Merge(size_t l, size_t mid, size_t h)
 {
-	int i = l, j = mid + 1;
+	size_t i = l, j = mid + 1;
 	dynamic_array<Collider*> A(16);
 	while (i <= mid && j <= h) {
 		if (list->get(i)->pos->x < list->get(j)->pos->x)
@@ -78,8 +78,8 @@ void MergeSort<Collider*>::Merge(int l, int mid, int h)
 		A.push_back(list->get(j));
 	}
 	//copy to the list
-	int k = 0;
-	for (int i = l; i <= h; i++)
+	size_t k = 0;
+	for (size_t i = l; i <= h; i++)
 	{
 		list->get(i) = A.get(k++);
 	}
