+++ dynamic_array.h
@@ -7,7 +7,7 @@ public:
 
 	~dynamic_array();
 	dynamic_array(size_t size);
-	T& operator[](int);
+	T& operator[](size_t);
 
 
 	void push_back(const T& item);
@@ -19,12 +19,12 @@ public:
 	int find(const T& item);
 
 	bool contains(const T& item);
-	void insert(int index, const T& item);
+	void insert(size_t index, const T& item);
 
 	void prepend(const T& item);
 	//removes all instances
 	void remove(const T& item);
-	void removeAtIndex(const int& index);
+	void removeAtIndex(const size_t& index);
 	void removeAll();
 
 	//gives the number of items it can hold
@@ -32,8 +32,8 @@ public:
 	//give the number of the actual items that is holding
 	size_t getCount();
 	void print();
-	void print(int& i, int& j);
-	T& get(int const& i);
+	void print(size_t& i, size_t& j);
+	T& get(size_t const& i);
 
 
 private:
